Implement CameraSimulationDataProvider
