+++ XRTK-Core/Packages/com.xrtk.core/Definitions/InputSystem/Simulation/MixedRealityCameraInputSimulationDataProviderProfile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7393bff647614a449a110d62bca0b16b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
