+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/InputSystem/Simulation/MixedRealityCameraInputSimulationDataProviderProfileInspector.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 94dd522e01e6c9c4fb24c876c5705a6a
+guid: a827998827def7942ae18e6974391627
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
+  icon: {instanceID: 0}
   userData: 
   assetBundleName: 
   assetBundleVariant:
