+++ XRTK-Core/Packages/com.xrtk.core/Inspectors/Profiles/InputSystem/Simulation/MixedRealityInputSimulationProfileInspector.cs
@@ -12,23 +12,6 @@ namespace XRTK.Inspectors.Profiles.InputSystem.Simulation
     [CustomEditor(typeof(MixedRealityInputSimulationProfile))]
     public class MixedRealityInputSimulationProfileInspector : BaseMixedRealityProfileInspector
     {
-        private SerializedProperty isCameraControlEnabled;
-
-        private SerializedProperty extraMouseSensitivityScale;
-        private SerializedProperty defaultMouseSensitivity;
-        private SerializedProperty mouseLookButton;
-        private SerializedProperty isControllerLookInverted;
-        private SerializedProperty currentControlMode;
-        private SerializedProperty fastControlKey;
-        private SerializedProperty controlSlowSpeed;
-        private SerializedProperty controlFastSpeed;
-        private SerializedProperty moveHorizontal;
-        private SerializedProperty moveVertical;
-        private SerializedProperty mouseX;
-        private SerializedProperty mouseY;
-        private SerializedProperty lookHorizontal;
-        private SerializedProperty lookVertical;
-
         private SerializedProperty handSimulationMode;
 
         private SerializedProperty simulateEyePosition;
@@ -66,23 +49,6 @@ namespace XRTK.Inspectors.Profiles.InputSystem.Simulation
         protected override void OnEnable()
         {
             base.OnEnable();
-
-            isCameraControlEnabled = serializedObject.FindProperty("isCameraControlEnabled");
-
-            extraMouseSensitivityScale = serializedObject.FindProperty("extraMouseSensitivityScale");
-            defaultMouseSensitivity = serializedObject.FindProperty("defaultMouseSensitivity");
-            mouseLookButton = serializedObject.FindProperty("mouseLookButton");
-            isControllerLookInverted = serializedObject.FindProperty("isControllerLookInverted");
-            currentControlMode = serializedObject.FindProperty("currentControlMode");
-            fastControlKey = serializedObject.FindProperty("fastControlKey");
-            controlSlowSpeed = serializedObject.FindProperty("controlSlowSpeed");
-            controlFastSpeed = serializedObject.FindProperty("controlFastSpeed");
-            moveHorizontal = serializedObject.FindProperty("moveHorizontal");
-            moveVertical = serializedObject.FindProperty("moveVertical");
-            mouseX = serializedObject.FindProperty("mouseX");
-            mouseY = serializedObject.FindProperty("mouseY");
-            lookHorizontal = serializedObject.FindProperty("lookHorizontal");
-            lookVertical = serializedObject.FindProperty("lookVertical");
 
             handSimulationMode = serializedObject.FindProperty("handSimulationMode");
 
@@ -130,31 +96,6 @@ namespace XRTK.Inspectors.Profiles.InputSystem.Simulation
 
             serializedObject.Update();
 
-
-            EditorGUILayout.PropertyField(isCameraControlEnabled);
-            {
-                EditorGUILayout.BeginVertical("Label");
-
-                EditorGUILayout.PropertyField(extraMouseSensitivityScale);
-                EditorGUILayout.PropertyField(defaultMouseSensitivity);
-                EditorGUILayout.PropertyField(mouseLookButton);
-                EditorGUILayout.PropertyField(isControllerLookInverted);
-                EditorGUILayout.PropertyField(currentControlMode);
-                EditorGUILayout.PropertyField(fastControlKey);
-                EditorGUILayout.PropertyField(controlSlowSpeed);
-                EditorGUILayout.PropertyField(controlFastSpeed);
-                EditorGUILayout.PropertyField(moveHorizontal);
-                EditorGUILayout.PropertyField(moveVertical);
-                EditorGUILayout.PropertyField(mouseX);
-                EditorGUILayout.PropertyField(mouseY);
-                EditorGUILayout.PropertyField(lookHorizontal);
-                EditorGUILayout.PropertyField(lookVertical);
-
-                EditorGUILayout.EndVertical();
-
-            }
-
-            EditorGUILayout.Space();
             EditorGUILayout.PropertyField(simulateEyePosition);
 
             EditorGUILayout.Space();
