+++ XRTK-Core/Packages/com.xrtk.core/Services/MixedRealityToolkit.cs
@@ -16,6 +16,7 @@ using XRTK.Interfaces.Diagnostics;
 using XRTK.Interfaces.InputSystem;
 using XRTK.Interfaces.NetworkingSystem;
 using XRTK.Interfaces.Providers.Controllers;
+using XRTK.Interfaces.Providers.InputSystem.Simulation;
 using XRTK.Interfaces.Providers.SpatialObservers;
 using XRTK.Interfaces.SpatialAwarenessSystem;
 using XRTK.Interfaces.TeleportSystem;
@@ -381,6 +382,22 @@ namespace XRTK.Services
                                     controllerDataProvider.Profile))
                                 {
                                     Debug.LogError($"Failed to start {controllerDataProvider.DataProviderName}!");
+                                }
+                            }
+                        }
+
+                        if (ActiveProfile.InputSystemProfile.InputSimulationDataProvidersProfile != null)
+                        {
+                            foreach (var inputSimulationDataProvider in ActiveProfile.InputSystemProfile.InputSimulationDataProvidersProfile.RegisteredInputSimulationDataProviders)
+                            {
+                                if (!CreateAndRegisterService<IMixedRealityInputSimulationDataProvider>(
+                                    inputSimulationDataProvider.DataProviderType,
+                                    inputSimulationDataProvider.RuntimePlatform,
+                                    inputSimulationDataProvider.DataProviderName,
+                                    inputSimulationDataProvider.Priority,
+                                    inputSimulationDataProvider.Profile))
+                                {
+                                    Debug.LogError($"Failed to start {inputSimulationDataProvider.DataProviderName}!");
                                 }
                             }
                         }
